Update mdate.cmd
Had to update the for loop of the mp4 files to for /f "tokens=*" %%f in ('dir /b .\*.mp4') do ( ... ) so that Windows 11 didn't pick up the newly created MP4 files and essentially loop forever. I posted https://stackoverflow.com/questions/75945252/windows-11-batch-cmd-file-for-loop-behavior-has-changed about this and a comment said this is how looping always works. I disagree, otherwise I wouldn't have been able to use this code for the past few years.